Adjusted various loop timers and control status checks
Adding setting for how quickly increment gets called when in continuous mode while an increment/decrement input is held. Added 1ms timers in device_detection() and bind(), which seems to massively reduce CPU load. Moved thread running and thread waiting checks and variable assignments to only happen in controls() so that it should be impossible for controls() to call increment() or switch() more than once in quick succession when it should only be called once